+++ components/src/core/user-menu/user-menu.component.ts
@@ -52,7 +52,7 @@ import { map, startWith } from 'rxjs/operators';
         </pxb-user-menu-avatar>
 
         <ng-template #menu>
-            <pxb-drawer-header
+            <pxb-user-menu-header
                 *ngIf="menuTitle"
                 class="pxb-user-menu-header"
                 [title]="menuTitle"
@@ -67,7 +67,7 @@ import { map, startWith } from 'rxjs/operators';
                 >
                     <ng-content select="[pxb-menu-avatar]"></ng-content>
                 </pxb-user-menu-avatar>
-            </pxb-drawer-header>
+            </pxb-user-menu-header>
             <ng-content select="[pxb-menu-header]"></ng-content>
             <ng-content select="[pxb-menu-body]"></ng-content>
         </ng-template>